push everything to limit

+++ instrumentation/src/ds_dpst.h
@@ -62,7 +62,7 @@ class DisjointSet
     // a map from finish dpst node id to finish
     robin_hood::unordered_map<int, hclib_finish*> all_finishes;
 
-    robin_hood::unordered_map<double, unsigned int> cache;
+    robin_hood::unordered_map<unsigned long int, unsigned int> cache;
     // boost::unordered_map<unsigned int, unsigned int> cache;
     // google::dense_hash_map<unsigned int, unsigned int, hash<unsigned int>, eqint> cache = google::dense_hash_map<unsigned int, unsigned int, hash<unsigned int>, eqint>(100000);
     // unordered_map<unsigned int,unsigned int> cache;
@@ -143,8 +143,8 @@ public:
     void print_table();
 
     // reachability queries, return if step_a preceds step_b
-    bool precede(tree_node_cpp* step_a, tree_node_cpp* step_b, int task_a, int task_b);
-    bool visit(tree_node_cpp* step_a, tree_node_cpp* step_b, int task_a, int task_b, robin_hood::unordered_set<int> &visited);
+    bool precede(tree_node_cpp* step_a, tree_node_cpp* step_b, unsigned int task_a, unsigned int task_b);
+    bool visit(tree_node_cpp* step_a, tree_node_cpp* step_b, unsigned int task_a, unsigned int task_b, robin_hood::unordered_set<int> &visited);
 
     tree_node_cpp* find_lca_left_child_cpp(tree_node_cpp* node1, tree_node_cpp* node2);
